pyproject.toml
[tool.poetry]
name = "linkorg-backend"
version = "0.1.0"
description = ""
authors = ["Andriy Orehov <andriyorehov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
mongoengine = "^0.18.2"
flask_cors = "^3.0"
marshmallow = "^3.2"
python-slugify = "^3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
